# openapi.yaml
openapi: "3.0.0"

info:
  title: Rahvaalgatus
  version: "1"
  description: |
    The public API of <https://rahvaalgatus.ee>.  
    Source code available at <https://github.com/rahvaalgatus/rahvaalgatus>.

    Don't forget to set the `Accept` header as otherwise you'll be getting HTML
    back.  
    You can also do client-side requests with JavaScript as the API responses
    have a `Access-Control-Allow-Origin: *` header.

servers:
  - url: "https://rahvaalgatus.ee"

tags:
  - name: Statistics
  - name: Initiatives

paths:
  /statistics:
    get:
      tags: [Statistics]
      summary: Get general statistics about Rahvaalgatus.
      description: |
        Note that the returned initiative counts do not at this time include
        external initiatives. External initiatives were the ones sent to the
        parliament through means other than Rahvaalgatus, and were imported to
        Rahvaalgatus via the Parliament API.

      responses:
        200:
          description: OK
          content:
            application/vnd.rahvaalgatus.statistics+json; v=1:
              schema: {$ref: "#/components/schemas/Statistics"}

  /initiatives:
    get:
      tags: [Initiatives]
      summary: Get a list of initiatives matching your filter.

      parameters:
        - name: phase
          in: query
          description: Filter initiatives by phase.
          schema:
            type: string
            example: "sign"
            enum: [edit, sign, parliament, government, done]

        - name: signedSince
          in: query
          description: |
            Limit returned initiatives to those that were signed since the
            given date.

            The date is interpreted to be in the "Europe/Tallinn" time zone. If
            you specify a full time and date with an offset, that will be
            honored.

          schema:
            type: string
            format: date
            example: "2020-01-01"

        - name: order
          in: query
          description: |
            Order initiatives. Ascending by default. Use the "-" prefix
            to request a descending sort.

          schema:
            type: string
            example: "-signatureCount"
            enum:
              - "signatureCount"
              - "-signatureCount"
              - "signaturesSinceCount"
              - "-signaturesSinceCount"

        - name: limit
          in: query
          description: Limit returned initiatives.
          schema:
            type: integer
            example: 5

      responses:
        200:
          description: OK
          content:
            application/vnd.rahvaalgatus.initiative+json; v=1:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Initiative"

        400:
          description: |
            Invalid request

            It could be because of an invalid phase or an invalid order query
            parameter.

  /initiative-events:
    get:
      tags: [Initiatives]
      summary: Get a list of initiative events matching your filter.

      parameters:
        - name: order
          in: query
          description: |
            Order initiative events. Ascending by default. Use the "-" prefix
            to request a descending sort.

          schema:
            type: string
            example: "-occurredAt"
            enum:
              - "occurredAt"
              - "-occurredAt"

        - name: limit
          in: query
          description: Limit returned initiative events.
          schema:
            type: integer
            example: 5

      responses:
        200:
          description: OK
          content:
            application/vnd.rahvaalgatus.initiative-event+json; v=1:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Initiative Event"

        400:
          description: |
            Invalid request

            It could be because of an invalid order query parameter.

  /initiatives/{initiativeUuid}:
    get:
      tags: [Initiatives]
      summary: Get the initiative's data.

      parameters:
        - name: initiativeUuid
          in: path
          description: Initiative's UUID
          required: true
          schema:
            type: string
            format: uuid
            example: "92cc16ee-107e-4208-b92c-2ffed24d4f4b"

      responses:
        200:
          description: OK
          content:
            application/vnd.rahvaalgatus.initiative+json; v=1:
              schema: {$ref: "#/components/schemas/Initiative"}

        404: {description: "Initiative not found."}

components:
  schemas:
    Statistics:
      type: object

      properties:
        initiativeCountsByPhase:
          type: object
          description: The count of initiatives by phase.

          properties:
            edit:
              type: integer
              example: 3
              description: |
                The number of initiatives in the edit phase.  
                These are shown as discussions on the web.

            sign:
              type: integer
              example: 1
              description: |
                The number of initiatives in the sign phase.  
                These are shown as being signable on the web.

            parliament:
              type: integer
              example: 3
              description: The number of initiatives in the parliament phase.

            government:
              type: integer
              example: 3
              description: The number of initiatives in the government phase.

            done:
              type: integer
              example: 7
              description: The number of initiatives in the government phase.

        signatureCount:
          type: integer
          example: 1336
          description: |
            The total number of digital signatures on all initiatives.

    Initiative:
      type: object

      properties:
        id:
          type: string
          format: uuid
          example: 92cc16ee-107e-4208-b92c-2ffed24d4f4b
          description: |
            The initiative's id.  
            This is a UUID as of Feb 20, 2020, but will become an integer in
            the next few months.

        title:
          type: string
          description: The initiative's title.
          example: Teeme elu paremaks!

        phase:
          type: string
          enum: [edit, sign, parliament, government, done]
          description: The phase of the initiative.
          example: "sign"

        signatureCount:
          type: integer
          nullable: true
          example: 1336
          description: |
            The digital signature count for the initiative.  
            Note that for external initiatives (those not originated from
            Rahvaalgatus) this may be null, indicating we don't know the count.

        signaturesSinceCount:
          type: integer
          nullable: true
          example: 5
          description: |
            The digital signature count for the initiative since the given date
            _if_ `signedSince` was specified in the query.  
            Will be missing if no `signedSince` is present.

    Initiative Event:
      type: object

      properties:
        id:
          type: integer
          example: 1
          description: |
            The event's id.  
            Note that for some events the id is a string. Such is the case for
            the "sent-to-parliament" event, for example.

        initiativeId:
          type: string
          format: uuid
          example: 92cc16ee-107e-4208-b92c-2ffed24d4f4b
          description: |
            The event's initiative's id.  
            This is a UUID as of Feb 20, 2020, but will become an integer in
            the next few months.

        title:
          type: string
          example: "Algatus jõudis Riigikokku"
          description: |
            The title of the event.  
            These will only be in Estonian at the moment, but localization is
            theoretically possible.

        occurredAt:
          type: string
          format: date-time
          example: "2015-06-18T13:37:42.666Z"
          description: |
            The time the event happened.  
            Note that this doesn't necessarily mean the event entered the
            Rahvaalgatus' database at that time. Rahvaalgatus is unfortunately
            often notified of parliamentary events days later, when they reach
            the [Riigikogu's API](https://api.riigikogu.ee).